[FIX] arg amount quote & create args

+++ include/utils.h
@@ -17,7 +17,7 @@ size_t strccount(const char *str, char c);
 char *get_arg(char *str, int nb);
 char *get_arg_quote(char *str, int nb);
 size_t arg_amount(char *str);
-size_t arg_amount_quote(char *str);
+int arg_amount_quote(char *str);
 char *read_bytes_starting_arg(char *str, int nb, size_t bytes);
 char *capitalize_cmd(char *cmd);
 int limit_nb(int nb, int max);
